insightface/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
insightface
numpy
scikit-learn
python-multipart
onnxruntime>=1.17.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
